[bezier-curve] Use static_cast in bernstein_polynomial() instead of implicit conversion

+++ src/vulkan-renderer/bezier_curve.cpp
@@ -18,7 +18,7 @@ std::uint32_t BezierCurve::binomial_coefficient(std::uint32_t n, const std::uint
 
 float BezierCurve::bernstein_polynomial(std::uint32_t n, std::uint32_t k, const float curve_precision,
                                         const float coordinate_value) {
-    return binomial_coefficient(n, k) * static_cast<float>(pow(curve_precision, k)) *
+    return static_cast<float>(binomial_coefficient(n, k)) * static_cast<float>(pow(curve_precision, k)) *
            static_cast<float>(pow(1 - curve_precision, n - k)) * coordinate_value;
 }
